core/cmake: alpaquita: change default lib64 to lib
Alpine used to apply its own downstream patch to fix a default value of
${CMAKE_INSTALL_LIBDIR} in cmake.  This patch was introduced in commit
a654934063 ("main/cmake: change default lib64 to lib").  Then this issue was
fixed on the cmake side [1] and the downstream patch was removed in commit
661c3dd489 ("main/cmake: upgrade to 3.21.0").  But it doesn't work for
Alpaquita, so we have to provide our own patch to fix this issue.

Without this change, abuild may fail to build packages that use cmake and
include module GNUInstallDirs, for example:

  >>> libsndfile*: Running postcheck for libsndfile
  >>> ERROR: libsndfile*: Packages must not put anything under /usr/lib64, use /usr/lib instead
  >>> ERROR: libsndfile*: prepare_subpackages failed
  >>> ERROR: libsndfile: rootpkg failed

[1] https://gitlab.kitware.com/cmake/cmake/-/commit/37b24e0bedcb2be97323d119897413addcb692e2

Conflicts:
	core/cmake/APKBUILD