core/lvm2: upgrade to 2.03.17
[ Alpine commit e599de8c22c7489eddf26ed439234d2315c2074a ]

re-add use_mlockall patch.

The memorylocking implementation without mlockall was introduced in
https://sourceware.org/git/?p=lvm2.git;a=commit;h=18b82048e457d6ff426d9dd828475d4dd02b8204

This was an attempt to reduce locking of "way too much memory (>100MB)
due to the way glibc works. I believe this is less fo an issue with musl
libc so use the stupid but simple mlockall by default to be safe.