core/apk-tools: do not check if 'bin/sh' is available
Because of this patch

  0002-Don-t-run-triggers-till-bin-sh-is-available.patch

apk checks if '/bin/sh' is available before run a package script.  If
this is false, apk doesn't treat this as an error and just skips the
script.

Since a busybox-binsh subpackage was added, this behaviour started
failing installation of busybox from scratch:

  + apk --root /chroot --initdb --update-cache --no-progress add alpaquita-base
  ...
  (7/25) Installing busybox (1.35.0-r27)
  Executing busybox-1.35.0-r27.post-install
  WARNING: busybox-1.35.0-r27.post-install: won't be executed, there's no /bin/sh yet
  busybox-1.35.0-r27.post-install: NOTE: consider use of --no-scripts or --no-commit-hooks options
  (8/25) Installing busybox-binsh (1.35.0-r27)
  (9/25) Installing alpaquita-baselayout (3.3.0-r1)
  Executing alpaquita-baselayout-3.3.0-r1.pre-install
  Executing alpaquita-baselayout-3.3.0-r1.post-install
  lib/apk/exec/alpaquita-baselayout-3.3.0-r1.post-install: line 36: chown: not found
  lib/apk/exec/alpaquita-baselayout-3.3.0-r1.post-install: line 37: chmod: not found
  ...

Even though 'bin/sh' is not available, it's not a problem, because
busybox uses shebang '#!/bin/busybox sh' in its scripts now.

Basically, this check only needs for installation of glibc-utils from
scratch (when busybox is not installed).  It's not been relevant since
commit dbf15e4445 ("core/glibc: don't depend on /bin/sh in the
trigger"), so let's remove the patch.