core/python3: check if avx objects exist before linking
There can be a situation when extra objects are passed to link() and
they may not have avx alternatives. For example, python bindings for
libseccomp fail to be built as follows:

  ld: cannot find ../.libs/libseccomp.a.avx2: No such file or directory

To fix this, check in advance if avx objects exist in a similar way as
it's done with '-l' options. Also, reduce code duplication.