core/binutils: add -z text by default
[ commit a737d101cad8c0acf782d9a0ec38cf8d3432b6ff ]

makes the linker fail on textrels directly.

we already check for textrels with scanelf- but this crashes in the
link that generates them, which makes it a bit easier to debug why they
occur.

it does, however, need an additional workaround to permit when it cannot
be fixed (adding -Wl,-z,notext to ldflags) on top of options=textrels