core/linux-headers: implement rsyncless install
[ commit faa03fc95331caeeb8ff8a558c6d237394010b06 ]

this is a bit of a hack- however it drops the dependency on `rsync` to
install the headers.

since they are just a tree of `.h` (and we drop the other files by
default), the output is the same.

since rsync depends on a lot of things, and linux-headers is a base
dependency for a lot of things, this eases the bootstrap dependency
chain.

in this case, it allows zstd to use a different build system without
causing a dependency loop[0]; but keeping file-only installations like this
more dependency-free is always nice.

[0]: rsync -> zstd -> meson -> python -> linux-headers -> rsync