universe/mandown: cargo fetch with --target
[ commit 39f6fef7137b3a09a6412f948123d9c588bb03c7 ]

this does not download deps that won't be used for the default build
target (the ctarget we are building for implicitly).
this makes quite a few fetches faster.. because a ton of crates download
a 10mb "windows-sys" that is never used, amongst other things.

maybe breaks a thing or two.