universe/polkit: abstract libs even more
[ commit be6d26add243b31e62e8a17407e3350c0c12fdf9 ]

presently, there is an issue: elogind-libs provide -libs.. which means
by default, things that depend on the libs can pull in either.

this sounds correct, and polkit-elogind is not affected (it depends on
polkit-elogind-libs), but there is an unseen issue:

polkit depends on polkit-libs- so it can pull in polkit-elogind-libs
even for base polkit. this mismatch is incorrect.

this additional layer of abstraction and replaces= placement means that
doing:

$ apk add polkit !polkit-elogind
$ apk add polkit-elogind

or whatever other fun stuff, will always leave a polkit with the correct
matching libs, and no libraries missing from disk (post-uninstall).

also drop bash (unneeded), move depends to package() to make clear it's
base polkit.

and, of course the virtual -libs means things depending on -libs for
whatever reason (nothing in aports/pmaports) will still resolve. should
be no breaking change :)