core/grub2: ignore new ext2 flags
After an update of e2fsprogs to 1.47.0 mkfs.ext4 creates
ext4 file systems with the metadata_csum_seed feature.
The problem is that grub-probe fails to detect such file
systems:

  ~ grub-probe -t fs -d /dev/sdb
  grub-probe: error: unknown filesystem.

The issue is already solved in the upstream grub repository.

In addition to that we are picking up an upstream grub patch
which makes grub ignore another new flag (large_dir), as it is
likely be required for the next update of e2fsprogs.