universe/go: skip misc/cgo/test on musl
Upstream GO marked such tests to be skipped on Alpine. For example:

https://github.com/golang/go/commit/42105eb8a5b59e715573f5299218e12be2dbb050

As their check by presence of /etc/alpine-release (and we don't
have this file), the tests are not actually skipped in our case.

A good way to address this issue would be to update the upstream Go
code and use a distro-neutral check for Musl. However, as it's
stated here:

https://git.musl-libc.org/cgit/musl/commit/?id=2e5fff43dd7fc808197744c67cca7908ac19bb4f

the next Musl release (1.2.4) is expected to make these tests pass
with Musl. So it seems rational to explicitly skip them in APKBUILD
and revisit the issue once we have the next Musl release.