core/libtasn1: drop broken patch
[ commit 1bbe496e91a18009491b42a72d93d8feae795d09 ]

musl's new behaviour in
https://github.com/bminor/musl/commit/b50eb8c36c20f967bd0ed70c0b0db38a450886ba
is compliant with what is expected for strverscmp,
so the inverted logic becomes wrong

technically speaking, the patch here papered over a difference in what the code did-
with the correct behaviour, "UNKNOWN" returns NULL. but before that
musl change, it returned the version- the opposite of what was really
documented. this test patch made it 'pass when the version is returned'

since this is very niche, nothing was broken in the end anyway.