universe/abuild: add -fno-plt for x86*
[ commit 9c497fa68472a5efa3e66d138917b91e1f01b03e ]

per https://lists.alpinelinux.org/~alpine/devel/%3C1628515011.zujvcn248v.none%40localhost%3E

this is a safe for-x86-only optimisation that decreases code size and
increases speed by around 1-3%

a small gain, but a 'free' one. the only incompatibility is it must be
disabled when -Wl,-z,lazy is set.

however, we generally force -z,now (immediate binding), so this is not a
fundamental incompatibility, and is usable for us

(if this seriously breaks something, then shoot me, but this will most likely be fine :) )