universe/go: remove cgo files from misc/
[ commit 480ebb7dfd3764a43bc26a291448c8669546d146 ]

Starting with >= 3.11.5 abuild complains about the size of the
go-doc package. This is due to the size of cgo related files in
misc/. These files will be moved elsewhere with Go 1.21 and do
not seem to be needed really (since we haven't shipped them with
the origin package). Hence we can just remove them for now to
significantly reduce the go-doc package size.