core/eudev: add sgx group post-install
[ commit 4f83ae597bfe4e7cbaa558687ebfa6fb7d30f232 ]

3.2.12 contains rules that assigns sgx devices to the sgx group, so a
warning is printed that the group is not recognised if it is not added
to the group list.

we don't ship sgx in our kernels however, so this is just a warning with
no meaning, so we don't need any guarantee of this group existing (i.e.
a post-upgrade migration)

ref #15043