core/cryptsetup: fixup man pages which use the .so directive
[ commit 85eaa116dbedfad3a70ad6da0cf5dbf05dacfba1 ]

The crpytsetup-doc package ships several man page which just consist
of an `.so` directive for referring to another man page. This directive
is not supported by mandoc [1]. As such, opening man pages such as
`cryptsetup-create(8)` does not work presently.

The recommended solution by upstream is to replace such man pages with
a symlink. This commits adds a shell oneliner which does exactly that.

[1]: https://inbox.vuxu.org/mandoc-tech/20101024194129.GJ20876@iris.usta.de/#r