universe/zeromq: build with CURVE security support again
[ commit f87ef4d699c4aab6b2d2b8b67d86e6006a18be07 ]

Support for CURVE security was silently dropped with the upgrade to
version 4.3.5 as upstream changed the default of ENABLE_CURVE as part of
this release [1]. Since the 4.3.5 release, zeromq is therefore also not
linked against libsodium anymore but as libsodium is still in
$makedepends I believe this to be unintentional.

Also: py3-pyzmq fails check() without this which is how this was
discovered. Alternatively, we could also patch py3-pyzmq to not
assume that libzmq provides CURVE security support.

[1]: https://github.com/zeromq/libzmq/commit/ff47aeb791e134a78bc386e12eea67618e0bf2f7#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL264