universe/dotnet6-runtime: kludge - add libunwind to makedepends
abuild automatically traces shared library dependencies.  That works
when a package gets built against libfoo-dev because that also pulls
in libfoo.  A newly built binary that dependens on libfoo.so gets the
dependency on libfoo detected and recorded.

Here we repackage previoulsy built binaries, so libunwind package that
provides libunwind.so is not necessarily installed (b/c we did not
need libunwind-dev to build the binaries now, we used it when we
actually built the binaries).

As a kludge add libunwind in makedepends to get the so dependency
detected and recorded.