core/kmod: remove kmod-static-nodes init.d service
[ commit 6a41588a798bb9a89e03926760b31a13201c02ec ]

The kmod-static-nodes init.d service only creates a file in the
/run/tmpfiles.d/ directory. However tmpfiles.d functionality is a
SystemD feature and the OpenRC compatible equivalent opentmpfiles
https://github.com/OpenRC/opentmpfiles is retired/no longer supported
and is not packaged for Alpine.

As there is no mechanism in Alpine to handle tmpfiles it therefore
makes no sense for kmod to have a init.d service/kmod-openrc
subpackage simply for creating a tmpfile that is never used.