core/openssh: remove sftp-interactive.patch
[ commit b544dbe9982060824869f64028919ccf17054542 ]

This patch changes code that is only executed when OpenSSH is compiled
without editline support. I assume, in 2014 (when this patch was added)
this was still the case. However, nowadays we compile OpenSSH with
editline support, hence the prompt is drawn by editline and we no longer
need this patch.

Nonetheless, since the problem still persists when OpenSSH is compiled
without editline support, I also proposed the patch for upstream
integration: https://github.com/openssh/openssh-portable/pull/480