core/xz: updowngrade to 5.6.1_p525 built from 5.2.5 sources
NB: This package is still called 5.6.1 to make apk think it's an
upgrade, but the contents of this package is downgraded to be built
from v5.2.5 as obtained with git archive from git.tukaani.org repo.

We also rename the library to liblzma.so.5.6.1 so that the libs
package provides so:liblzma.so.5=5.6.1 Otherwise apk gets confused
when a package that has a greater version provides an .so that has
a lesser version.