core/dbus: remove unnecessary dependency
[ commit 9af813daea04a24b37d14f2da2ba7f6dd60fa0e3 ]

dbus does not require `util-linux`.

I compared the build log when building with and without util-linux and
there are no meaningful differences (some files are compiled in non
deterministic order, so they shuffled by a line or two, but nothing of
consequence).

Maybe it was required in 2011, when this dependency was introduced.

See: a6e77729e40b32c980615aff653faa0b155495fe