universe/rust: allow re-running build()
[ commit 157db54c2b021afa15ea139ce5061721fca8cc85 ]

This comes in handy when debugging why a build fails,
use `--incremental` so we do not have to rebuild all
stages each time build() is run.

Also, `./configure` will exit with the following message
if a "config.toml" (even an empty one) is found:
"configure: ERROR: Existing 'config.toml' detected. Exiting"

So, if "config.toml" exists, assume we want to restart the
build by skipping `./configure` and proceeding on to `./x.py`.