core/openssh: fix build with BOOTSTRAP=1 on glibc
openssh falls back to use of utmpx functions if the login() function is not
available or explicitly disabled at a configure stage with the
'--disable-libutil' option.  openssh is supposed to be built without utmpx
support if BOOTSTRAP=1.  This is true for musl, but glibc provides utmpx
functionality out of the box.  This leads to the following error:

  loginrec.c: In function 'wtmpx_write':
  loginrec.c:1287:18: error: '_PATH_WTMP' undeclared (first use in this function)
   1287 |         updwtmpx(WTMPX_FILE, utx);
        |                  ^~~~~~~~~~
  loginrec.c:1287:18: note: each undeclared identifier is reported only once for each function it appears in
  loginrec.c: In function 'wtmpx_get_entry':
  loginrec.c:1357:24: error: '_PATH_WTMP' undeclared (first use in this function)
   1357 |         if ((fd = open(WTMPX_FILE, O_RDONLY)) < 0) {
        |                        ^~~~~~~~~~

The reason is this patch

  avoid-redefined-warnings-when-building-with-utmps.patch

explicitly undeclares '_PATH_WTMP'.

Let's not apply this patch if we are bootstrapping openssh on glibc.

Fixes: aadf67bd25b ("core/openssh: fix building with gcc14 on glibc")