core/glib: suppress output of glib-compile-schemas trigger
[ commit c78330c32f104a5db14e3f0759f7fa4e46484442 ]

The output of compile-schemas, such as warnings about bad dconf paths,
is not intended to be read by end users and is meant only for developers.

Supress only these warnings if /bin/sh is a symlink to busybox.
Otherwise discard all warnings, for example if the user has dash-binsh
or yash-binsh installed.