core/execline and others: fix (wrong) linter warnings
[ commit 62e566e6a1c471f64d1456c9aae71ae238ed3a09 ]

 lint warns for the use of braces in ${_skalibs_version}.
It is true that they're unnecessary, but they help with clarity,
which is not a bad thing, and the syntax is entirely normal when
writing shell scripts. That warning should not be there.
Especially when it triggers at random, in some places and not
others (and sometimes it picks at random *on the same line*).

 It also warns for `backticks` instead of $(this), which is a
little bit more legit because $(this) allows for better nesting,
but `backticks` are, again, perfectly supported shell constructs
and in the absence of nesting there's no reason not to use them.

 lint should be fixed.

Signed-off-by: Laurent Bercot <ska@appnovation.com>