core/docbook-xsl: make docbook-xsl depend on both -ns and -nons
[ commit 50bfb70b3d9884cd0c8ba51a47470aafac335d7e ]

It seems that many packages still use the old URI
http://docbook.sourceforge.net/release/xsl and for some reason don't mind
that we previously pointed it to a different variant of the XSL templates
(or fail silently as PostgreSQL used to?). Since this URI is now registered
for docbook-xsl-nons and docbook-xsl installs only docbook-xsl-ns, it broke
the build of many packages.

They can be installed at the same time, so there's no reason not to.

Related to #16579