core/busybox: enable cut -F
[ commit 79fddc2aa9c64c47fb3da6edeae9570b4e9caa7e ]

cut -f is limited to only single character delimiters which means it's not
usable to handle the output of most shell commands.

cut -F is often times a simpler alternative to awk. When the number of
fields is unknown it can be a major simplification compared to awk.

As far as I have seen this doesn't cause a size increase cause the code
has not been optimized out by the compiler.