core/file: Import patch file instead of downloading it from GitHub
[ commit bd570c2a8fc5988357b1415c93c0c69b3f4c8996 ]

As noted by Celeste at [0], it's in general not recommended to use URL-based
patches. Back when the patch was merged, that was deemed OK for 'master' though.
As the download of the patch file is now indeed failing due to a changed
checksum, use the opportunity to import the patch into the repo (as was done on
the stable branches via e89c62d8 for example).

Please note that the checksum has changed because the patch was downloaded with
the '?full_index=1' option.

[0]: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/77814#note_470256