core/distro-base: Tweak /etc/issue to avoid incorrect English
[ commit 5931d4c5bda95bc6272b3e0ed3875f0b3b045b8b ]

Currently the message displayed above the login prompt on virtual
consoles looks like this:

  Welcome to Alpine Linux 3.21
  Kernel 6.12.8-0-virt on an x86_64 (/dev/ttyS0)

The "on an $arch" part results in correct English for some of
the supported architectures (e.g. x86_64 and aarch64) but not
for some others (e.g. riscv64 and loongarch64).

Tweak the message so that correct English is produced regardless
of the architecture.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>