core/apk-tools: pin to libapk2
In Alpaquita there are multiple versions of apk-tools present
in our apk repos, so there are several packages providing
`so:libapk.so.2.14.9`. As a result, when we try to call
`apk fetch --recursive` it's possible that apk-tools of
a previous revision is fetched instead of libapk2:

    # apk fetch --recursive apk-tools
    Downloading apk-tools-2.14.9-r0
    Downloading apk-tools-2.14.9-r1
    Downloading ca-certificates-bundle-20241121-r1
    Downloading glibc-2.39.0-r1
    Downloading libcrypto3-3.3.3-r0
    Downloading libssl3-3.3.3-r0
    Downloading zlib-1.3.1-r2
    #

Pinning subpackages to the same libapk2 revision solves this issue.