core/curl: add --with-ca-path to curl build
[ commit 5c35191091b85b1d10b96de568e5824d02a670e3 ]

Prior to !69012, curl automatically set --with-ca-bundle and
--with-ca-path to /etc/ssl/cert.pem and /etc/ssl/certs, respectively.
However, now if --with-ca-bundle alone is set, then --with-ca-path
will not be set. If an incomplete or wrong CA bundle were provided to
curl, curl would not be able to validate the root certificate even if
the ca-certificates package were installed. This commit effectively
restores the previous behavior by explicitly setting --with-ca-path.

fixes #16980