core/openrc: fix 0102-openrc-add-rc_quiet_openrc-for-openrc-binary.patch
Starting from openrc-0.60 the patch is applied in a wrong way.  The
reason is rc_conf_yesno("rc_quiet_openrc") is done before configuration
files are processed.  This is not a big deal since rc_conf_yesno()
invokes rc_conf_value() which in turn processes the configuration files
if this has not been done before.  However the following condition:

  errno == ENOENT

becomes true even though the 'rc_quiet_openrc' option has been found.
The reason is there are many cases where the 'errno' variable can be set
to 'ENOENT' when rc_conf_yesno() processes the configuration files.
This leads to that 'rc_quiet_openrc=NO' specified in a configuration
file is ignored.

To fix this, put the code after the configuration files are processed.
Also, add a clarification comment.