universe/alsa-{lib,utils}: remove update_config_sub
In alpine, if the config.sub doesn't support loongarch64, they use
update_config_sub function to update the config.sub for a newer version.
To check whether it supports loongarch64, they execute the config.sub
with argument loongarch64 and if it exits with non-zero status they
update the config.sub.
In this case, this is false positive because config.sub exits with 127,
since it cannot find the shebang interpreter(config.sub has /usr/bin/sh
and alpine doesn't have that file, instead they have /bin/sh).
But the config.sub is new enough to support loongarch64, hence the
removal of update_config_sub from prepare.