core/grub: never use amove internal globbing
[ commit 06471599904835737e652e4fec73336ce91e3e55 ]

The effect of unescaped globbing is (in my opinion) more intuitive to
understand than the fact that (at least for now) escaped globbing will
still be globbed inside of amove. Since !87698 using 'amove *' reliably
does the globbing before invoking amove which will make it possible to
drop the globbing inside of amove later on. That should allow easier
to understand, safer and smaller code in abuild and APKBUILDs.