universe/mesa: set -mtls-dialect=gnu for armhf&armv7
[ commit c5b2299328038a20a7405da059d76c7ca3aa59e4 ]

With gcc the configure step otherwise sets -mtls-dialect=gnu2, which
causes mesa to stop working on arm after the upgrade to mesa-25.

The issue was bisected to commit 44bda7c2 ("dri: put shared-glapi into
libgallium.*.so") in mesa, and the bug reported in
https://gitlab.freedesktop.org/mesa/mesa/-/issues/13236

Alpine used to set the same flag for armhf&armv7 before the upgrade to
mesa-24.2.1.