universe/rust: fix llvm-config for cross compilation correctly
[ commit f226c1074eb9b05f18ace962eedc95c9b1062e16 ]

for some reason target.*.llvm-config needs to be an absolute path and
target.$_target.llvm-config needs to be changed as well. Sorry for not
testing the patch correctly.

Note: The rust buildsystem expects the llvm-config from CBUILD to work
for CHOST too (same prefix and more). When this is not the case one
needs to create a custom llvm-config script for CHOST.

Fixes 7abccbd05b09 main/rust: fix llvm-config not found on non-alpine systems