universe/rust: wrap llvm-config for CHOST
[ commit 6f53dfd3196920ff809b3416255ae3243903580f ]

Upstream intends using CBUILD llvm-config for CHOST. This means that
it needs to be the same llvm version installed to the same prefix with
the same components and all compiler flags need to be compatible with
the cross compiler. All of these might not be true but fixing all of
these is too complicated.

Upstream workarounds include replacing CBUILD with CHOST in some paths
but this doesn't work for abuild.

This APKBUILD only prepends CBUILDROOT to include/link/bin dir and
corrects components by passing --sysroot (added to $CFLAGS by abuild)
and wrapping llvm-config. This is only enought to cross compile on
alpine.

For other cases try changing the wrapper to run the correct (CHOST)
llvm-config with qemu or similar. Or work on a proper fix like
https://github.com/llvm/llvm-project/issues/9777