universe/rust: use stage1 artifacts at packaging stage when bootstrapping
For some reason, there are no required artifacts in stage2 directories
after bootstrap build.  But they can be found in stage1 directories.  As
long as they are not from a pre-compiled stage 0, I think it's okay to
use them since a bootstrapped rust is needed only as an intermediate
build.