core/fontconfig: re-enable tests on glibc
The test for "Basic functionality with the bind-mounted cache dir"
that is cited as the reason is not even run by default as bubblewrap
package required for that test is not listed in checkdepends.  (But
with bubblewrap in place the test does succeed).

(cherry picked from commit 99abe73d9be23e61fdf611b6fcfd8d7d0d3a891e)