universe/go: skip tsan_test on musl
In commit be392aad58b ("core/gcc: enable libsanitizer on x86_64"),
libsanitizer was enabled on musl/x86_64.  This enabled libsanitizer
related tests in Go including tsan_test which contains a number of Go
programs each of them fails as follows:

  $ setarch -R go run -v cmd/cgo/internal/testsanitizers/testdata/tsan.go
  ThreadSanitizer: CHECK failed: tsan_platform_linux.cpp:488 "((thr_beg)) >= ((tls_addr))" (0x7ffff5b26380, 0x7ffff5b26af8) (tid=2107)
      #0 <null> <null> (libtsan.so.2+0x8b3b4) (BuildId: 5f4f15be1723dc0b0a1815bda48a12dd574cd893)
      #1 <null> <null> (libtsan.so.2+0xc9b3c) (BuildId: 5f4f15be1723dc0b0a1815bda48a12dd574cd893)
      #2 <null> <null> (libtsan.so.2+0x8925e) (BuildId: 5f4f15be1723dc0b0a1815bda48a12dd574cd893)
      #3 <null> <null> (libtsan.so.2+0xa6f1a) (BuildId: 5f4f15be1723dc0b0a1815bda48a12dd574cd893)
      #4 <null> <null> (libtsan.so.2+0x3daef) (BuildId: 5f4f15be1723dc0b0a1815bda48a12dd574cd893)
      #5 <null> <null> (ld-musl-x86_64.so.1+0x6ace6) (BuildId: cc4437768727e38e703f6120e6bca0b15cd7aa0b)

  exit status 66

In Alpine, this test was skipped because it requires 'setarch' which
requires priviledged access to containers [1].  This is not fully true.
This test also can be run without 'setarch' and there is a chance it
will pass (as it currently passes on glibc).  Moreover, this test also
fails with 'setarch' on musl and looks like the problem is on the
libsanitizer side.

[1] https://gitlab.alpinelinux.org/alpine/aports/-/commit/5613387dbb1