core/busybox: adjust action_type in `struct init_action`
[ commit c792a0618f1b628cd3eac4e5af83f5ce0bf00caa ]

Commit b45665c4d121480c23d4dc426cd1214b43af8ab2 added support for
the reboot action to BusyBox init, as part of this a new reboot
bitmask was introduced and the restart bitmask value was incremented
to 0x100. Therefore, the action can no longer be represented in a
8-bit unsigned integer and was changed to uint32_t in most parts
of the code with the aforementioned commit.

However, there was a small oversight as the action_type in `struct
init_action` was not updated to be a uint32_t. Therefore, the restart
bitmask was no longer checked correctly. This commit fixes this
oversight.

Fixes #17447