core/distro-baselayout: document filesystem hierarchy
[ commit cfd87286a0fcec5efb7682badeda474ddf27a823 ]

Include a man page describing the current filesystem hierarchy on
Alpine. This document describes current practices on Alpine (i.e.: it
seeks to be descriptive rather than prescriptive).

It loosely aligns with the following (at least in the portions that
apply to Alpine itself):

- https://man7.org/linux/man-pages/man7/hier.7.html
- https://man.openbsd.org/hier
- https://refspecs.linuxfoundation.org/FHS_3.0/fhs/index.html
- https://www.freedesktop.org/software/systemd/man/latest/file-hierarchy.html
- https://uapi-group.org/specifications/specs/linux_file_system_hierarchy/

The man page is written in mdoc(7), which includes semantic markup for
paths, links etc. This should produce higher-quality conversion to other
formats (e.g.: HTML) than the legacy man(7) format.

Fixes: https://gitlab.alpinelinux.org/alpine/aports/-/issues/16426