core/java-common: fix missing symlink when the jvm is empty
[ commit e1cd3a1d1057f981527e93cc83213c778f49bd62 ]

In case a jvm is uninstalled, its folder keep some empty folders. This
can trick the trigger script to link it as "default-jvm", and produce
no /usr/bin/java symlink at the end.

Checking that the bin/java file is there seems enough to be sure the jvm
folder is sane enough.

[ BS: added the check for bin/java to the find command ]