universe/podman: patch for stopping rootless podman containers
[ commit c720b2564302371e26b7080615a4ce6a8d128476 ]

Attempting to stop a rootless Podman container currently results in the
error Error: given PID did not die within timeout. This problem was also
faced by users of Chimera Linux, who implemented a fix and discussed
further in their GitHub issue

  https://github.com/chimera-linux/cports/issues/1718

This patch is simple and I have tested it as working on my workstation.