core/busybox: silence grep error in default udhcpc script
[ commit e0e709e5b3aa2aaeb9db35590bfd2aedd7f0882e ]

There is no guarantee that $RESOLV_CONF exists already. In case this is
the first time we are a generating it, or if is stored on a tmpfs (e.g.,
when using testing/captive-browser), then grep will emit an error
message that the file could not be found. Since this is somewhat
surprising error message to receive on bootup, this commit silences all
errors emitted by grep.

This is a fixup for c22bfc406d86e32f863d308ea72a5d0e93c254a2 (CC: @omni).