core/apk-tools: Fix scripts not working with apk --root
We experience the same problem with `apk --root` vs minirootfs as
described in this issue:

  https://gitlab.alpinelinux.org/alpine/apk-tools/-/issues/11161

so as a conservative interim measure apply a local patch that reverts
the commit from right before 3.0.0 that introduced the problem and get
back to the 3.0.0_rc9 behavior.

To be dropped when the fix comes from upstream.